Split provider fallback challenge tests
